Fixed current tests, adding power verification
